chore(deps): use ora as spinner

+++ src/main.ts
@@ -1,26 +1,25 @@
 #! /usr/bin/env node
 
 import { existsSync, readFileSync, writeFileSync } from "node:fs";
+import { writeFile } from "node:fs/promises";
 import { resolve } from "node:path";
 
 import cac from "cac";
-import { createSpinner } from "nanospinner";
-import color from "picocolors";
-
 import { detect } from "detect-package-manager";
+import ora from "ora";
+import color from "picocolors";
 
 import { Logger } from "./logger";
 
 import { installVitest, removeJestDeps } from "./deps";
 import { formatSetupFile, formatVitestConfig } from "./formatter";
-import { getTestFiles } from "./fs";
 import { getJestConfig } from "./jest";
 import { transformJestConfigToVitest } from "./mapper";
-import { transformJestScriptsToVitest } from "./scripts";
-import { constructDOMCleanupFile } from "./setup";
-import { transformJestTestToVitest } from "./transformer";
 
-import type { CleanupFile } from "./setup";
+import { getTestFiles } from "./fs";
+import { transformJestScriptsToVitest } from "./scripts";
+import { type CleanupFile, constructDOMCleanupFile } from "./setup";
+import { transformJestTestToVitest } from "./transformer";
 
 const cli = cac();
 cli
@@ -35,58 +34,64 @@ cli
   });
 
 const args = cli.parse();
-// force new line
-const spinner = createSpinner(color.green("Finding Jest config...\n"), {
-  color: "green",
-}).start();
+const spinner = ora(color.green("Finding Jest config...\n"));
 
 Logger.init(args.options.debug);
 
 try {
+  spinner.start();
+
   const { path, config } = await getJestConfig();
 
   Logger.debug(
     path.length
       ? `Using Jest configuration from ${path}`
-      : "Jest configuration not found. Using default settings.",
+      : "Jest configuration not found. Using default configuration values.",
   );
 
-  spinner.update({
-    text: color.green("Configuring Vitest based on Jest configuration...\n"),
-    color: "green",
-  });
+  spinner.text = color.green(
+    "Configuring Vitest based on Jest configuration...\n",
+  );
 
-  const vitestConfig = transformJestConfigToVitest(config, args.options.globals);
+  const vitestConfig = transformJestConfigToVitest(
+    config,
+    args.options.globals,
+  );
 
   let setupFile: CleanupFile | undefined;
-  let isTS = true;
+  const isTS = existsSync(resolve(process.cwd(), "tsconfig.json"));
 
   Logger.debug("Vitest configuration generated");
 
-  spinner.update({
-    text: color.green("Transforming test files...\n"),
-    color: "green",
-  });
+  spinner.text = color.green("Scanning directory for test files...\n");
 
-  // TODO: write transformer script here...
   const testFiles = await getTestFiles(vitestConfig);
-  const transformedTests = transformJestTestToVitest(testFiles, args.options.globals);
 
-  for (const transformedTest of transformedTests) {
-    Logger.debug(`Successfully transformed ${transformedTest.path}`);
-    writeFileSync(transformedTest.path, transformedTest.content);
+  if (args.options.dryRun) {
+    Logger.info(`${testFiles.length} test files found.`);
+  } else {
+    spinner.text = color.green(
+      `Transforming ${testFiles.length} test file(s)...\n`,
+    );
+
+    const transformedTests = transformJestTestToVitest(
+      testFiles,
+      args.options.globals,
+    );
+
+    await Promise.all(
+      transformedTests.map((test) => writeFile(test.path, test.content)),
+    );
+
+    Logger.debug(`Succesfully transformed ${testFiles.length} test file(s)`);
   }
 
-  Logger.debug(`Succesfully transformed ${testFiles.length} test file(s)`);
-
-  spinner.update({
-    text: color.green("Transforming package's scripts...\n"),
-    color: "green",
-  });
   const packageJsonPath = resolve(process.cwd(), "package.json");
 
-  if (existsSync(packageJsonPath)) {
+  if (existsSync(packageJsonPath) && !args.options.dryRun) {
     Logger.debug("package.json found. Jest scripts will be transformed.");
+
+    spinner.text = color.green("Transforming package's scripts...\n");
 
     const packageManager = await detect();
     const packageJson = JSON.parse(readFileSync(packageJsonPath).toString());
@@ -94,10 +99,6 @@ try {
       ...Object.keys(packageJson.dependencies),
       ...Object.keys(packageJson.devDependencies),
     ];
-
-    if (!dependencies.includes("typescript")) {
-      isTS = false;
-    }
 
     const newScripts = transformJestScriptsToVitest(packageJson.scripts);
     packageJson.scripts = newScripts;
@@ -108,10 +109,7 @@ try {
 
     setupFile = constructDOMCleanupFile(vitestConfig, dependencies);
 
-    spinner.update({
-      text: color.green("Performing dependency cleanup...\n"),
-      color: "green",
-    });
+    spinner.text = color.green("Performing dependency cleanup...\n");
 
     installVitest(packageManager, dependencies);
 
@@ -128,53 +126,54 @@ try {
     );
   }
 
-  spinner.update({
-    text: color.green("Writing Vitest config (and setup files)...\n"),
-    color: "green",
-  });
+  if (args.options.dryRun) {
+    spinner.text = "✨ Dry-run completed successfully.\n";
+  } else {
+    spinner.text = color.green("Writing Vitest config (and setup files)...\n");
 
-  if (setupFile) {
-    const setupText = formatSetupFile(setupFile);
-    const setupFilename = `vitest.setup.${isTS ? "ts" : "js"}`;
+    if (setupFile) {
+      const setupText = formatSetupFile(setupFile);
+      const setupFilename = `vitest.setup.${isTS ? "ts" : "js"}`;
 
-    if (Array.isArray(vitestConfig?.setupFiles)) {
-      vitestConfig.setupFiles.push(`./${setupFilename}`);
+      if (Array.isArray(vitestConfig?.setupFiles)) {
+        vitestConfig.setupFiles.push(`./${setupFilename}`);
+      }
+
+      const setupFilepath = resolve(process.cwd(), setupFilename);
+
+      writeFileSync(setupFilepath, setupText);
+
+      Logger.debug(
+        `Successfully written Vitest setup file on ${setupFilepath}`,
+      );
+    } else {
+      Logger.debug(
+        "Setup file is not necessary as target environment is not DOM.",
+      );
     }
 
-    const setupFilepath = resolve(process.cwd(), setupFilename);
+    const configText = formatVitestConfig(vitestConfig);
+    const configFilename = resolve(
+      process.cwd(),
+      `vitest.config.${isTS ? "ts" : "js"}`,
+    );
 
-    writeFileSync(setupFilepath, setupText);
+    writeFileSync(configFilename, configText);
 
-    Logger.debug(`Successfully written Vitest setup file on ${setupFilepath}`);
-  } else {
     Logger.debug(
-      "Setup file is not necessary as target environment is not DOM.",
+      `Successfully written Vitest configuration file on ${configFilename}`,
+    );
+
+    spinner.succeed(
+      color.green(
+        `✨ Succesfully converted Jest test suite to Vitest. You're good to Vitest 🚀\n`,
+      ),
     );
   }
-
-  const configText = formatVitestConfig(vitestConfig);
-  const configFilename = resolve(
-    process.cwd(),
-    `vitest.config.${isTS ? "ts" : "js"}`,
-  );
-
-  writeFileSync(configFilename, configText);
-
-  Logger.debug(
-    `Successfully written Vitest configuration file on ${configFilename}`,
-  );
-
-  spinner.success({
-    text: color.green(
-      `✨ Succesfully converted Jest test suite to Vitest. You're good to Vitest 🚀\n`,
-    ),
-  });
 } catch (err) {
   const error = err as Error;
 
-  spinner.error({
-    text: color.red(`❌ Transformation failed due to ${error.message}\n`),
-  });
+  spinner.fail(color.red(`❌ Transformation failed due to ${error.message}\n`));
 
   Logger.error(error);
 }
